bonkh/requirements.txt
SQLAlchemy==1.1.13
ba_colander==1.0b1
deform==2.0.7
pyramid==1.10.1
